Change tooltip for control points.

+++ src/main/java/com/simona/model/ControlPoint.java
@@ -17,6 +17,6 @@ public class ControlPoint {
     private Station station;
 
     private Integer stn_sys_id;
-    private Integer freq;
+    private Double freq;
     private Integer status;
 }
